# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BetRoyal Betting & Casino Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/me/kyc:
    post:
      operationId: submitKyc
      tags: [users]
      summary: Submit KYC documents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycStatus"
  /users/me/change-password:
    post:
      operationId: changePassword
      tags: [users]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /wallet/deposit:
    post:
      operationId: createDeposit
      tags: [wallet]
      summary: Create a deposit request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /wallet/withdraw:
    post:
      operationId: createWithdrawal
      tags: [wallet]
      summary: Create a withdrawal request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /wallet/transactions:
    get:
      operationId: getTransactions
      tags: [wallet]
      summary: Get transaction history
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
...
  /games:
    get:
      operationId: getGames
      tags: [games]
      summary: Get all games
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: provider
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameList"
...
  /games/{id}:
    get:
      operationId: getGame
      tags: [games]
      summary: Get game details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
  /games/{id}/launch:
    post:
      operationId: launchGame
      tags: [games]
      summary: Launch a game session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameLaunchResponse"
...
  /games/{id}/favorite:
    post:
      operationId: toggleFavoriteGame
      tags: [games]
      summary: Toggle game favorite
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # BETS
...
  /admin/users/{id}/wallet:
    post:
      operationId: adminAdjustWallet
      tags: [admin]
      summary: Manually adjust user wallet (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletAdjustInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
...
  /admin/games/{id}:
    patch:
      operationId: updateAdminGame
      tags: [admin]
      summary: Update game settings (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminGameUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # AUTH
...
    User:
      type: object
      required: [id, username, email, role, status, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin, agent]
        status:
          type: string
          enum: [active, suspended, banned]
        avatar:
          type: ["string", "null"]
        kycStatus:
          type: string
          enum: [none, pending, approved, rejected]
        vipLevel:
          type: integer
        referralCode:
          type: string
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        username:
          type: string
        avatar:
          type: string
    KycInput:
      type: object
      required: [fullName, documentType, documentNumber]
      properties:
        fullName:
          type: string
        documentType:
          type: string
        documentNumber:
          type: string
        documentFront:
          type: string
        documentBack:
          type: string
    KycStatus:
      type: object
      required: [status, message]
      properties:
        status:
          type: string
        message:
          type: string
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6
...
    Wallet:
      type: object
      required: [id, userId, balance, bonusBalance, currency]
      properties:
        id:
          type: integer
        userId:
          type: integer
        balance:
          type: number
        bonusBalance:
          type: number
        currency:
          type: string
...
    DepositInput:
      type: object
      required: [amount, method]
      properties:
        amount:
          type: number
          minimum: 1
        method:
          type: string
        currency:
          type: string
    WithdrawalInput:
      type: object
      required: [amount, method, address]
      properties:
        amount:
          type: number
          minimum: 10
        method:
          type: string
        address:
          type: string
    Transaction:
      type: object
      required: [id, userId, type, amount, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [deposit, withdrawal, bet, win, bonus, referral]
        amount:
          type: number
        status:
          type: string
          enum: [pending, completed, failed, cancelled]
        method:
          type: ["string", "null"]
        reference:
          type: ["string", "null"]
        createdAt:
          type: string
    TransactionList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # GAMES
    Game:
      type: object
      required: [id, name, slug, category, provider, rtp, isActive, isFeatured]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        category:
          type: string
          enum: [slots, live_casino, sports, crash, mines, dice, plinko, limbo, keno, coin_flip, lucky_wheel, fishing, cards, blockchain]
        provider:
          type: string
        thumbnail:
          type: ["string", "null"]
        rtp:
          type: number
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        isPopular:
          type: boolean
        maxWin:
          type: ["number", "null"]
        minBet:
          type: number
        maxBet:
          type: number
        isFavorited:
          type: boolean
        playCount:
          type: integer
    GameList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Game"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    GameLaunchResponse:
      type: object
      required: [sessionId, gameUrl]
      properties:
        sessionId:
          type: string
        gameUrl:
          type: string

    # BETS
...
    WalletAdjustInput:
      type: object
      required: [amount, type, reason]
      properties:
        amount:
          type: number
        type:
          type: string
          enum: [credit, debit]
        reason:
          type: string
    AdminGameUpdate:
      type: object
      properties:
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        isPopular:
          type: boolean
        rtp:
          type: number
        minBet:
          type: number
        maxBet:
          type: number
...